module5-task1 [refractoring] (script)  made more readability structure
