+++ js/data.js
@@ -1,0 +1,28 @@
+export const getData = () => {
+  const DESCRIPTION_FOR_IMAGE_POST = [
+    'Солнечный закат на горизонте',
+    'Закат над озером',
+    'Пейзаж с горами и деревьями',
+    'Зимний пейзаж с снегом',
+    'Портрет человека на фоне города',
+    'Птицы на ветке',
+    'Небоскрёбы в ночи',
+    'Лес в осеннем озере',
+    'Городской пейзаж в тумане',
+    'Картина с цветами',
+    'Пустынный пейзаж',
+    'Старинный замок',
+    'Лагуна на пляже',
+    'Лесной огонь',
+  ];
+
+  const MESSAGES_FOR_IMAGE_POST = ['Всё отлично!', 'В целом всё неплохо.', 'Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', 'В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают.', 'Как можно было поймать такой неудачный момент?!'];
+
+  const NAMES_FOR_IMAGE_POST = ['Макар', 'Адам', 'Богдан', 'Платон', 'Леон', 'Савелий', 'Есения', 'Эмилия', 'Агата', 'Аделина', 'Ясмина', 'Оливия', 'Аврора', 'Мира', 'Агния', 'Амалия'];
+
+  return {DESCRIPTION_FOR_IMAGE_POST, MESSAGES_FOR_IMAGE_POST, NAMES_FOR_IMAGE_POST};
+};
